Update README and mapping
Slight tweaks to the README to include instructions for utilizing Poetry
shell as a virtualenv as the project utilizes it. This makes it simpler
for developers to get up and running with it should they choose to do
so.

Additionally added a slight tweak to the `process_creation` mapping
section as many clients with Sysmon may parse out the specific rule
name, which in this case is `ProcessCreate`.