Updates pyproject and removes newlines for piping
Updates the pyproject file to get pySigma to a more recent version as
the version was considerably behind. Additionally slightly tweaks
`cli/convert.py` to remove the newline entries. Users can introduce
newlines in their terminals to improve readability. By introducing the
characters in the query output it may introduce issues when users pipe
input to `pbcopy`, `xsel`, `xclip` or the like.