Update mapping.yaml for common Windows operations
Updates to various Windows operations in the `config/mapping.yml`, including a
duplicate of the change made in [PR #1](https://github.com/Securonix/SigmaToSecuronix/pull/1)
for the `process_creation` events.

Depending on the parsing that clients use the `deviceAction` can be
parsed out as the rule name, so this config change (FileCreate,
DnsQuery, etc.) allows for those options to
exist in the generated Snypr queries.

These could later be extended to filter for `deviceEventCategory` which
more accurately encompass these events.